refactor(app): App クラスを別ファイルで切り分け
